% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fancyCorr_function.R
\name{fancyCorr}
\alias{fancyCorr}
\title{Function for creating nicely-formatted correlation matrices.}
\usage{
fancyCorr(
  data,
  digits = 2,
  p.digits = 3,
  listwise = TRUE,
  p.adj = FALSE,
  sd.on.diag = TRUE,
  p.on.upper = TRUE,
  stars = FALSE,
  star.thresholds = c(0.05, 0.01, 0.001),
  method = "pearson"
)
}
\arguments{
\item{data}{Data frame containing the variables to be included in the matrix.}

\item{digits}{The number of decimals to which the correlation coefficeents and
standard deviations (if applicable) should be printed. Defaults to 2.}

\item{p.digits}{The number of decimals to which the p-values should be printed. Defaults to 3.}

\item{listwise}{Logical; should all rows containing missing data be removed? Defaults to TRUE.}

\item{p.adj}{Logical; should the p-values be adjusted for multiple comparisons? Defaults to FALSE.}

\item{sd.on.diag}{Logical; should the diagonal be replaced with standard deviations? Defaults to TRUE.}

\item{p.on.upper}{Should p-values for the correlations be printed on the upper triangle? Defaults to TRUE.}

\item{stars}{Should statistically significant correlation coefficients be starred? Defaults to FALSE.}

\item{star.thresholds}{A vector of three values corresponding to p-value thresholds that should
receive one, two, or three stars. Only operative if stars=TRUE. Defaults to c(.05, .01, .001).}

\item{method}{method="pearson" is the default value. The alternatives to be passed to cor are "spearman" and "kendall"}
}
\description{
\code{fancyCorr} creates nicely-formatted correlation matrices.
}
\details{
This function creates correlation matrices. Created as a mercy to my PSYC 5410 students.
}
\examples{
set.seed(110)
data <- data.frame(matrix(rnorm(50), ncol=5))
fancyCorr(data)
fancyCorr(data, digits=3, p.digits=3, listwise=FALSE, sd.on.diag=FALSE,
  p.adj=FALSE, stars=TRUE, star.thresholds=c(0,0,.1), p.on.upper=FALSE)

}
